Add amplitude event and code value to all code copy (#11)
### Summary
This PR overrides the original book.js file to directly send an
Amplitude event on code copy (line 696). The change to book.js is scoped
to adding the Amplitude event only to avoid regression and breaking
change.

It also cleans up previous added scripts to inject class into the copy
button, as they are no longer necessary.

### To Test
On clicking the copy button for any code snippet, an Amplitude event for
"CodeSnippetCopied" should be sent, with the code as its value.